+++ optuna/pruners/percentile.py
@@ -3,9 +3,9 @@ import math
 
 import numpy as np
 
+from optuna._trial_state import TrialState
 from optuna.pruners import BasePruner
 from optuna.study import StudyDirection
-from optuna.trial import TrialState
 from optuna import type_checking
 
 if type_checking.TYPE_CHECKING:
